Implement graceful termination for the executor
Graceful termination can allow us a much smoother operations workflow
on k8s with regards to a rolling executor restart. Instead of pausing
the executor and waiting for it to drain its jobs (which is only
really possible using monitoring systems) one can now use this in a
preStop hook and just delete the pod.

Change-Id: I20e29742c34dab5ccd2b939c40e87b6231176b8a