Don't decrease window size on merge failure
Dependent pipelines decrease the window size to limit the resource
impact that happens in case of a gate reset. Merge failures are
typically detected right after enqueue and are carried on until it's
the first item those don't really cause expensive gate resets. Thus
it's not necessary to decrease the window size if a merge failure
happens.

Change-Id: Icc8684c1bd8b5ce4e8eefcd0ee716ad9c771e689