GitHub Reporter: Fix User Email in Merge Commit Message
GitHub does not provide a review users email address via the
API pull request resources.
(cf. https://developer.github.com/v3/pulls/reviews/#list-reviews-for-a-pull-request)
Therefore, in its current implementation, the GitHub Reporter
always includes ``None`` as an email address in the ``Reviewed-by``
clause(s) of a merge commit message.

An email address of a GitHub user must be retrieved via the
user resource as done as per this change, so the GitHub Reporter
will add correct email addresses of reviewers.

Please note that if a user decides not to set a publicly visible
email address, the GibHub API will not include an email field for
the user object, and thus the ``Reviewed-by`` clause will still
state ``None`` as a users email in these cases.

Change-Id: I43e1fbe54e85c4daee18aae1eaf30bc00356d58c