Resume jobs after reenqueue of an item
Currently it can happen that a buildset cannot proceed when there is a
paused job and an in progress build is removed due to a
reconfiguration. The reason for this is that there will never be a
resume event for the paused job if there are no other in progress
builds.

This can be fixed by resuming jobs if necessary when the reenqueue is
finished.

Change-Id: Id470a8db310bb7833a7103b6fe287fd9b132e55f