Make test_playbook more stable
In some cases test_playbook fails on contended nodes. The reason is
that the job timeout repeatedly times out already during the pre
playbook causing it to re-run instead of failing with result
TIMED_OUT. To make this more robust give it a higher timeout so it can
reach the run playbook.

Change-Id: I042c25217138c3d96683b1494fc3be9cf395f6a1