Add an icon next to result buttons on the console log
As a low-cost attempt to indicate that the results button is
clickable and will bring up details, add a search-with-plus icon
("zoom in") to the left of it.  Let's see if that indicates to
users that they can click on it (or at least mouseover and read
the help text that explains it).

This also updates the other icons in the task rows to use the
newest recommendations for PF4.  There are some minor size and
style differences, so keeping all of them in sync makes sense.

Change-Id: I3fff295d2b3b38b9e1c75e13f9a8a967e812168e