Keep active nav links highlighted while browsing the page
So far, the highlighting of the active navigation item vanished when
clicking somewhere on the page. AFAIK this behaviour was unchanged with
PF4.

Fix this behaviour by using a NavLink component with the correct PF4 CSS
class for an active nav item.

The nav items will be highlighted as long as the user doesn't switch to
a sub page (e.g. /builds -> /build/12345). Once switched, the
highlighting will again vanish as the path doesn't match the route's
path rule anylonger (/build/12345 != /builds/*).

Change-Id: Id17de4086ee567703ba6c8cae124b131d3ef072c