Fix fetching function on build result page
It looks like this was never working because the function call supplied
to many parameters.

Fix by removing the additional and unnecessary null parameter.

I would keep this change after the PF4 changes of the buildset page
since it otherwise might result in some awkward behaviour due to the
different tabs and their contents.

Change-Id: I6fa62343a22d7a54e85a7e8d0babdabdc08b6352