GitHub Reporter: Fix User Email in Merge Commit Message
Re-introduce change https://review.opendev.org/#/c/738590/

There was a bug in fetching a GitHub user while loading pull requests.
`getUser` was given a project object instead of a project name string,
so no installation id was found for GitHub app authentication. That lead
to unauthenticated requests to the user API which resulted in HTTP 401
errors.

This change fixes this by passing the correct `project.name` to
`getUser` and also refactors the parameter names of the involved
methods from `project` to `project_name` to be more precise there.

This reverts commit 56f355ab72aab2716e2f8b580a242a2a824ea5a2.

Change-Id: Ib25f0ef5d27115626792890d0ec282ed0d562485