Don't mask exceptions in _process_event
For better logs we emit start and end messages before and after github
event processing. However the return of the event is part of the
finally clause which inhibits any exception from bubbling up. Moving
the return out of the finally clause fixes this so exceptions can be
caught by the caller and logged.

Change-Id: I25b40fe595eb149e0b52156aeceff9b464130d7f