Add reno configuration settings
This adds additional reno config to tell it to scan back to the earliest
versions when we started using reno to make sure we include all notes in
the output.

It also says not to ignore null merges so that we can merge temporary
branches back into the master branch, and reno will still proceed past
the null merge.

Change-Id: I75051cd492124441befc0dff5b33ccf0442b7754
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>