Move mark comparision into ZuulMark object
A ConfigurationError object makes a ConfigurationErrorKey from
<context, mark, error-string>.  The ConfigurationErrorKey assumes the
"mark" argument is a ZuulMark object that has fields such as "line"
and "snippet".

However there is a path to set config errors via
QueueItem.setConfigError(string) which makes a ConfigurationError with
<None, None, error-string>.  Consequently the "mark" object in the
ConfigurationErrorKey isn't actually a ZuulMark and the __eq__
comparision for ConfigurationErrorKey objects can fail.

This can hit in PipelineManager._findRelevantErrors() which walks the
error list and ends up doing a comparision with

  if ((err.key not in
       parent_layout.loading_errors.error_keys)

I think the result in production is that you can get a range of errors
that do not get reported.  I came to looking at this because I
accidentally added Ansible playbooks under .zuul.d and something was
clearly wrong, but I was not getting any error messages.

The solution here is to move the field matching for the ZuulMark into
an __eq__ method in the ZuulMark object itself, allowing None to be a
valid value.

Adding a file that looks like a playbook triggers this path for the
test_dynamic_split_config test, which passes with this fix.

Change-Id: If59f5556141db4bda54a61b459ffbffb935094ae