github status api docs : update
I was trying to understand the consequences of the various github
reporting APIs and what they can and can not report upstream.

This is a suggested edit on the existing docs for the check API, that
maybe makes things a bit clearer.  Some of the links are updated and a
couple of typos.

Change-Id: I03ff72029dae55714e51ca9f23735551e64b0d35