github: use the same status url for commit status as checks
We do a nice thing with github checks api where, when we report the
url for the check, we report the buildset URL if it's available,
otherwise the deep-link to the status page.  However, for the
non-checks-api commit status, we only send the URL of the overall
status page.  We didn't have a nice buildset page (or even a deep
link status page) at the time we wrote that, and we never upgraded.

Now that we've got all the details worked out for how to handle that
and produce various fallbacks, use the same method for both cases.

Change-Id: If5c5fa3c8b33b349d95eb52175e4faf17f5bbb36