Merge file comments from multiple tasks
When multiple tasks (or roles) return file_comments for the
same file via zuul_return only the last value is taken into account.

This uses a similar approach like for artifacts to merge the
file comments of multiple zuul_return calls. With that, all file
comments are merged additive into a final data structure. File comments
for the same file and/or the same line are kept distinct and don't
override each other.

Change-Id: Ib7300c243deaeb7922182a94cfa41e617fc4e6e6